+++ src/libponyc/codegen/codegen.h
@@ -23,7 +23,7 @@ PONY_EXTERN_C_BEGIN
 char* LLVMGetHostCPUName();
 void LLVMSetUnsafeAlgebra(LLVMValueRef inst);
 void LLVMSetReturnNoAlias(LLVMValueRef fun);
-void LLVMSetDereferenceable(LLVMValueRef fun, uint32_t i, uint64_t size);
+void LLVMSetDereferenceable(LLVMValueRef fun, uint32_t i, size_t size);
 
 // In case we need to change the internal calling convention.
 #define GEN_CALLCONV LLVMFastCallConv
